+++ shared/clients/model/src/traits/mod.rs
@@ -6,13 +6,11 @@
 pub mod focus;
 pub mod interpreter;
 pub mod layout;
-pub mod overlay;
 pub mod panel;
 
 pub use {
     focus::{Focus, FocusManager},
     interpreter::{DefaultLayoutInterpreter, LayoutInterpreter},
     layout::Layout,
-    overlay::{OverlayManager, OverlayRenderer},
     panel::Panel,
 };